:bug: 修复appsettings.json文件读取辅助类不能读取的问题
